feat(qemu): set RAM size for "virt" machine to 8GB
Booting a host with 3GB of RAM is constraining.  As such increase that
amount to 8GB, a number that is more realistic.  To support the increase
in RAM the amount of secure RAM reserved for the L1 GPT needs to be
bumped to 2MB.  The total size of the RMM image also needs to be bumped
by 16MB to account for the increased size of the .bss section.

Change-Id: I9e1a326f70e32b491ba462bb4c4a31bedf3c0adc
Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>
Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>