Allow TF-A launch RMM succesfully with MEASURED_BOOT
Thanks Bokdeuk Jeong for the investigation.

QEMU measured boot overwrites the TF-RMM segment with measured boot event log
data that's intended for the non-secure world.

In the commented out code the value of ns_log_addr is 0x40100000 which is the
exactly same address where RMM memory segments starts and there TF-A jumps to
actually start the RMM boot.

Signed-off-by: Lukasz Pawelczyk <l.pawelczyk@samsung.com>