Implement handling of additional SMCs (RMI) for sealing mechanism
This patch provides an implementation of additional Islet service
that the only purpose is to route additional vendor specific RMI
that allows to populate the realm metadata and a SMC for VHUK
retrieval. These additional SMCs are needed for the purpose of
the sealing key derivation mechanism.

Additionally, the patch provides implementation of VHUK retrieval
from the HES application using the PSA/serial interface.

Signed-off-by: Piotr Sawicki <p.sawicki2@samsung.com>